requirements.txt
python-telegram-bot>=8.1.1
requests>=2.18.1
redis>=2.10.6
